requirements.txt
# основные
Django==4.2.19
djangorestframework==3.15.2
django-filter==25.1
django-haystack==3.3.0
django-pandas==0.6.7
numpy==1.26.4
pandas==2.3.0
scikit-learn==1.6.1
scipy==1.13.1
surprise==1.1.4
torch==2.2.2
transformers==4.52.4
sentence-transformers==4.1.0
gensim==4.3.3
node2vec==0.5.0
Whoosh==2.7.4
scrapy==2.13.2
scrapy-playwright==0.0.43
scrapy-rotating-proxies==0.6.2
selenium==4.29.0
webdriver-manager==4.0.2
tqdm==4.67.1

# все установленные библиотеки полученные pip freeze > requirements.txt

# absl-py==2.3.0
# asgiref==3.8.1
# astunparse==1.6.3
# attrs==25.3.0
# Automat==24.8.1
# beautifulsoup4==4.13.3
# bs4==0.0.2
# cachetools==5.5.2
# certifi==2025.6.15
# cffi==1.17.1
# charset-normalizer==3.4.1
# click==8.1.8
# constantly==23.10.4
# cryptography==44.0.2
# cssselect==1.3.0
# dateparser==1.2.1
# defusedxml==0.7.1
# Django==4.2.19
# django-filter==25.1
# django-haystack==3.3.0
# django-pandas==0.6.7
# djangorestframework==3.15.2
# exceptiongroup==1.2.2
# filelock==3.18.0
# flatbuffers==1.12
# fsspec==2025.5.1
# fuzzywuzzy==0.18.0
# gast==0.4.0
# gensim==4.3.3
# google-auth==2.40.3
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# greenlet==3.2.3
# grpcio==1.73.0
# h11==0.14.0
# h5py==3.14.0
# hf-xet==1.1.5
# huggingface-hub==0.33.0
# hyperlink==21.0.0
# idna==3.10
# importlib_metadata==8.7.0
# incremental==24.7.2
# itemadapter==0.11.0
# itemloaders==1.3.2
# Jinja2==3.1.6
# jmespath==1.0.1
# joblib==1.5.1
# keras==2.9.0
# Keras-Preprocessing==1.1.2
# libclang==18.1.1
# lightgbm==4.6.0
# lxml==5.3.1
# Markdown==3.8.2
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# ml-dtypes==0.3.2
# mpmath==1.3.0
# namex==0.1.0
# networkx==3.2.1
# nltk==3.9.1
# node2vec==0.5.0
# numpy==1.26.4
# oauthlib==3.3.1
# opt_einsum==3.4.0
# optree==0.16.0
# outcome==1.3.0.post0
# packaging==24.2
# pandas==2.3.0
# parsel==1.10.0
# pillow==11.2.1
# playwright==1.52.0
# Protego==0.4.0
# protobuf==3.19.6
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
# pycparser==2.22
# PyDispatcher==2.0.7
# pyee==13.0.0
# Pygments==2.19.2
# pyOpenSSL==25.0.0
# PySocks==1.7.1
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# pytz==2025.2
# PyYAML==6.0.2
# queuelib==1.7.0
# RapidFuzz==3.13.0
# regex==2024.11.6
# requests==2.32.3
# requests-file==2.1.0
# requests-oauthlib==2.0.0
# rich==14.0.0
# rsa==4.9.1
# safetensors==0.5.3
# scikit-learn==1.6.1
# scikit-surprise==1.1.4
# scipy==1.13.1
# Scrapy==2.13.2
# scrapy-playwright==0.0.43
# scrapy-rotating-proxies==0.6.2
# selenium==4.29.0
# selenium-stealth==1.0.6
# sentence-transformers==4.1.0
# service-identity==24.2.0
# six==1.17.0
# smart-open==7.1.0
# sniffio==1.3.1
# sortedcontainers==2.4.0
# soupsieve==2.6
# sqlparse==0.5.3
# sympy==1.14.0
# tensorboard==2.9.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.16.2
# tensorflow-estimator==2.9.0
# tensorflow-io-gcs-filesystem==0.37.1
# tensorflow-macos==2.9.2
# termcolor==3.1.0
# threadpoolctl==3.6.0
# tldextract==5.1.3
# tokenizers==0.21.1
# tomli==2.2.1
# torch==2.2.2
# tqdm==4.67.1
# transformers==4.52.4
# trio==0.29.0
# trio-websocket==0.12.2
# Twisted==24.11.0
# typing==3.7.4.3
# typing_extensions==4.12.2
# tzdata==2025.2
# tzlocal==5.3.1
# undetected-chromedriver==3.5.5
# urllib3==2.3.0
# w3lib==2.3.1
# webdriver-manager==4.0.2
# websocket-client==1.8.0
# websockets==15.0.1
# Werkzeug==3.1.3
# Whoosh==2.7.4
# wrapt==1.17.2
# wsproto==1.2.0
# xgboost==2.1.4
# zipp==3.23.0
# zope.interface==7.2
